requirements.txt
accessible-pygments==0.0.4
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.0
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-widget-tweaks==1.4.12
docutils==0.19
environ==1.0
frozenlist==1.4.0
idna==3.4
imagesize==1.4.1
isodate==0.6.1
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
model-bakery==1.14.0
multidict==6.0.4
packaging==23.1
phonenumbers==8.13.18
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.6
pydata-sphinx-theme==0.13.3
Pygments==2.16.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-book-theme==1.0.1
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
zeep==4.2.1
